refactor: extract root.go god-file into internal/ packages (#173)
Split cmd/root.go (659 lines) into three focused internal packages and six
small cmd-package files, reducing root.go to 72 lines.

New internal packages (each with test file):
- internal/errs  — FmtAPIError + FromV2 (flag-independent error formatting)
- internal/client — Params/Get/Override/Reset (SDK client caching)
- internal/output — TableColumn/ResolveFormat/Print + all render helpers

New cmd-package files (thin wrappers, zero changes to resource files):
- cmd/errors.go  — apiErrFromV2 reads --debug flag, delegates to errs.FromV2
- cmd/client.go  — GetArubaClient loads config, delegates to client.Get
- cmd/output.go  — TableColumn type alias + resolveOutputFormat + PrintOutput
- cmd/messages.go — msgCreated/Updated/Deleted/Action/DryRun helpers
- cmd/prompt.go  — confirmDelete + readSecret
- cmd/helpers.go — projectRef/projectWrapper/listOpts/extractIDFromURI

Closes #173
