+++ cmd/helpers_coverage_test.go
@@ -353,71 +353,3 @@ func TestProjectWrapper(t *testing.T) {
 	})
 }
 
-// ─── printJSON / printYAML ────────────────────────────────────────────────────
-
-type fakeRawMarshaler struct{ j, y []byte }
-
-func (f fakeRawMarshaler) RawJSON() []byte { return f.j }
-func (f fakeRawMarshaler) RawYAML() []byte { return f.y }
-
-func TestPrintJSON(t *testing.T) {
-	t.Run("nil emits {}", func(t *testing.T) {
-		out := captureStdout(func() { printJSON(nil) })
-		if strings.TrimSpace(out) != "{}" {
-			t.Fatalf("got %q, want {}", out)
-		}
-	})
-
-	t.Run("rawMarshaler written verbatim", func(t *testing.T) {
-		payload := []byte(`{"id":"x1"}`)
-		out := captureStdout(func() { printJSON(fakeRawMarshaler{j: payload}) })
-		if !strings.Contains(out, `"id":"x1"`) {
-			t.Fatalf("got %q", out)
-		}
-	})
-
-	t.Run("rawMarshaler empty bytes emits {}", func(t *testing.T) {
-		out := captureStdout(func() { printJSON(fakeRawMarshaler{j: []byte{}}) })
-		if strings.TrimSpace(out) != "{}" {
-			t.Fatalf("got %q, want {}", out)
-		}
-	})
-
-	t.Run("plain struct marshalled", func(t *testing.T) {
-		out := captureStdout(func() { printJSON(map[string]any{"hello": "world"}) })
-		if !strings.Contains(out, `"hello"`) || !strings.Contains(out, `"world"`) {
-			t.Fatalf("got %q", out)
-		}
-	})
-}
-
-func TestPrintYAML(t *testing.T) {
-	t.Run("nil emits {}", func(t *testing.T) {
-		out := captureStdout(func() { printYAML(nil) })
-		if strings.TrimSpace(out) != "{}" {
-			t.Fatalf("got %q, want {}", out)
-		}
-	})
-
-	t.Run("rawMarshaler written verbatim", func(t *testing.T) {
-		payload := []byte("id: x1\n")
-		out := captureStdout(func() { printYAML(fakeRawMarshaler{y: payload}) })
-		if !strings.Contains(out, "id: x1") {
-			t.Fatalf("got %q", out)
-		}
-	})
-
-	t.Run("rawMarshaler empty bytes emits {}", func(t *testing.T) {
-		out := captureStdout(func() { printYAML(fakeRawMarshaler{y: []byte{}}) })
-		if strings.TrimSpace(out) != "{}" {
-			t.Fatalf("got %q, want {}", out)
-		}
-	})
-
-	t.Run("plain struct converted to YAML", func(t *testing.T) {
-		out := captureStdout(func() { printYAML(map[string]any{"hello": "world"}) })
-		if !strings.Contains(out, "hello: world") {
-			t.Fatalf("got %q", out)
-		}
-	})
-}
